feat: move lock to the core of the wasmd contract
