+++ precompiles/wasmd/wasmd.go
@@ -2,12 +2,14 @@ package wasmd
 
 import (
 	"embed"
+	"encoding/binary"
 	"fmt"
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/tracing"
 	"github.com/ethereum/go-ethereum/core/vm"
+	"github.com/ethereum/go-ethereum/crypto"
 
 	"cosmossdk.io/log"
 	storetypes "cosmossdk.io/store/types"
@@ -93,6 +95,15 @@ func (p Precompile) RequiredGas(input []byte) uint64 {
 
 // Run executes the wasmd precompile
 func (p Precompile) Run(evm *vm.EVM, contract *vm.Contract, readOnly bool) (bz []byte, err error) {
+	bz, err = p.run(evm, contract, readOnly)
+	if err != nil {
+		return cmn.ReturnRevertError(evm, err)
+	}
+	return bz, nil
+}
+
+// run executes the wasmd precompile (internal function)
+func (p Precompile) run(evm *vm.EVM, contract *vm.Contract, readOnly bool) (bz []byte, err error) {
 	// Initialize the context, db and chain data
 	ctx, stateDB, method, initialGas, args, err := p.RunSetup(evm, contract, readOnly, p.IsTransaction)
 	if err != nil {
@@ -101,6 +112,11 @@ func (p Precompile) Run(evm *vm.EVM, contract *vm.Contract, readOnly bool) (bz [
 
 	// This handles any out of gas errors
 	defer cmn.HandleGasError(ctx, contract, initialGas, &err)()
+
+	// Ensure the reentrancy lock
+	if err := p.ensureLock(evm.Origin, stateDB, method); err != nil {
+		return nil, err
+	}
 
 	// Now we call the method based on the function
 	switch method.Name {
@@ -147,4 +163,56 @@ func (Precompile) IsTransaction(method *abi.Method) bool {
 // Logger returns the logger for the precompile
 func (p Precompile) Logger(ctx sdk.Context) log.Logger {
 	return ctx.Logger().With("evm extension", "wasmd")
+}
+
+// ensureLock ensures that a reentrancy lock is set and not broken for the target contract
+// Reentrance lock is built using: precompile address, origin and origin nonce
+//   - Args are avoided to build the lock key, since the attacker may manipulate it
+//
+// This is done under a transient key under Cosmos SDK's stateDB
+// The lock is released at the end of the transaction by Cosmos SDK itself
+// Locks are specifically done per transaction and not per contract:
+// - Due to a limitation between EVM and WASM gas handling
+// - And to reduce the scope of the reentrances
+// - States generated as transient states are never committed and cleared at the end of the block
+func (p Precompile) ensureLock(
+	origin common.Address,
+	stateDB vm.StateDB,
+	method *abi.Method,
+) error {
+	// Build the lock key
+	lockKey := buildReentrancyLockKey(p.Address(), origin, stateDB.GetNonce(origin), method)
+
+	// Check if already locked
+	if stateDB.GetTransientState(p.Address(), lockKey) == common.BytesToHash([]byte{1}) {
+		return fmt.Errorf(
+			"reentrancy detected in precompile %s, method %s",
+			p.Address().Hex(), method.Name,
+		)
+	}
+
+	// Set lock
+	stateDB.SetTransientState(p.Address(), lockKey, common.BytesToHash([]byte{1}))
+	return nil
+}
+
+// buildReentrancyLockKey builds a deterministic lock key:
+//
+//	H( "wasmd.precompile.reentrancy.lock:", precompileAddr, originAddr, originNonce[:], and methodID )
+func buildReentrancyLockKey(
+	precompileAddr common.Address,
+	origin common.Address,
+	originNonce uint64,
+	method *abi.Method,
+) common.Hash {
+	var nonceBytes [8]byte
+	binary.BigEndian.PutUint64(nonceBytes[:], originNonce)
+
+	return crypto.Keccak256Hash(
+		reentrancyPerTargetNs,
+		precompileAddr.Bytes(),
+		origin.Bytes(),
+		nonceBytes[:],
+		method.ID,
+	)
 }
